message prover binary in e2e
