+++ crates/e2e/README.md
@@ -1,0 +1,7 @@
+# ZKISM E2E on Celestia
+
+## Generate a message proof payload using config
+
+```bash
+cargo run -p e2e --bin message-prover --release
+```
