+++ crates/e2e/src/config.rs
@@ -1,8 +1,8 @@
 #![allow(unused)]
 
-pub(crate) const MAILBOX_ADDRESS: &str = "0xb1c938f5ba4b3593377f399e12175e8db0c787ff";
-pub(crate) const MERKLE_TREE_ADDRESS: &str = "0xfcb1d485ef46344029d9e8a7925925e146b3430e";
+pub const MAILBOX_ADDRESS: &str = "0xb1c938f5ba4b3593377f399e12175e8db0c787ff";
+pub const MERKLE_TREE_ADDRESS: &str = "0xfcb1d485ef46344029d9e8a7925925e146b3430e";
 
-pub(crate) const EV_RPC: &str = "http://127.0.0.1:8545";
-pub(crate) const EV_WS: &str = "ws://127.0.0.1:8546";
-pub(crate) const TARGET_HEIGHT: u64 = 100;
+pub const EV_RPC: &str = "http://127.0.0.1:8545";
+pub const EV_WS: &str = "ws://127.0.0.1:8546";
+pub const TARGET_HEIGHT: u64 = 250;
